code-review-enterprise/ai-service/app/auth.py
"""Internal service-to-service JWT verification."""

from fastapi import HTTPException, Security
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from jose import JWTError, jwt
from app.config import settings

_bearer = HTTPBearer()


async def verify_internal_token(
    credentials: HTTPAuthorizationCredentials = Security(_bearer),
) -> str:
    """Verify the JWT issued by the Next.js API layer."""
    try:
        payload = jwt.decode(
            credentials.credentials,
            settings.JWT_SECRET,
            algorithms=[settings.JWT_ALGORITHM],
        )
        user_id: str = payload.get("id") or payload.get("sub")
        if not user_id:
            raise HTTPException(status_code=401, detail="Invalid token payload")
        return user_id
    except JWTError:
        raise HTTPException(status_code=401, detail="Invalid or expired token")
