code-review-enterprise/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.11.0
redis==5.0.8
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.6.0
pymongo==4.9.2
slowapi==0.1.9
